Rename `src/@types/ui.ts` to `ui-types.ts`
